Remove augeasconfigurator
